pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarket-pipeline"
version = "0.1.0"
description = "Postgres-first Polymarket Gamma + CLOB ingestion pipeline"
authors = [
    { name = "Your Name" }
]
requires-python = ">=3.10"

dependencies = [
    "psycopg[binary]>=3.1",
    "psycopg-pool>=3.2",
    "requests>=2.31",
    "sqlalchemy>=2.0",
    "pandas>=2.0",
    "python-dotenv>=1.0"
]

[project.scripts]
pm = "pm.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]